+++ lib/src/promotion_box/sbb_promotion_box_badge.dart
@@ -2,6 +2,10 @@ import 'package:flutter/material.dart';
 
 import '../../sbb_design_system_mobile.dart';
 
+const _shadowSpreadRadius = 8.0;
+const _borderRadius = 20.0;
+const _padding = EdgeInsets.symmetric(horizontal: 8.0, vertical: 1.0);
+
 /// Only used within the [SBBPromotionBox].
 class SBBPromotionBoxBadge extends StatelessWidget {
   const SBBPromotionBoxBadge({
@@ -18,16 +22,41 @@ class SBBPromotionBoxBadge extends StatelessWidget {
     final style = SBBControlStyles.of(context).promotionBox!;
     final resolvedBadgeColor = badgeColor ?? style.badgeColor;
     return Container(
-      padding: const EdgeInsets.symmetric(
-        horizontal: 8.0,
-        vertical: 1.0,
-      ),
+      padding: _padding,
       decoration: BoxDecoration(
         border: Border.all(color: style.borderColor!),
-        borderRadius: const BorderRadius.all(Radius.circular(20.0)),
+        borderRadius: BorderRadius.circular(_borderRadius),
         color: resolvedBadgeColor,
       ),
       child: Text(text, style: style.badgeTextStyle, maxLines: 1),
+    );
+  }
+}
+
+class SBBPromotionBoxBadgeShadow extends StatelessWidget {
+  const SBBPromotionBoxBadgeShadow({required this.badgeSize, this.shadowColor});
+
+  final Size badgeSize;
+  final Color? shadowColor;
+
+  @override
+  Widget build(BuildContext context) {
+    final style = SBBControlStyles.of(context).promotionBox!;
+    return Container(
+      decoration: BoxDecoration(
+        border: Border.all(
+          width: 0.0,
+          color: SBBColors.transparent,
+        ),
+        borderRadius: BorderRadius.circular(_borderRadius),
+        boxShadow: [
+          BoxShadow(
+            color: shadowColor ?? style.badgeShadowColor!,
+            spreadRadius: _shadowSpreadRadius,
+          ),
+        ],
+      ),
+      child: SizedBox.fromSize(size: badgeSize),
     );
   }
 }