+++ lib/src/shared/shared.dart
@@ -1,1 +1,1 @@
-export 'closeable_box_controller.dart';
+export 'closable_box_controller.dart';