+++ CHANGELOG.md
@@ -17,7 +17,9 @@ It is expected that you keep this format strictly, since we depend on it in our 
 ### Changed
 
 - BREAKING: default constructor of `SBBPromotionBox` is not `const`
-- BREAKING: removed `isCloseable` in `SBBPromotionBox` - merge behavior with `onClose`
+- BREAKING: removed `isCloseable` in `SBBPromotionBox` - merge behavior with nullable `onClose`
+- BREAKING: changed `description` to `subtitle` in `SBBPromotionBox` for more consistency with other Widgets
+- BREAKING: renamed `CloseableBoxController` to `ClosableBoxController`
 
 ## [2.3.0] - 2025-03-04